Move IPv4 test under IPv4.
