+++ test/unit-test/FreeRTOS_IP/FreeRTOS_IP_utest.c
@@ -1687,556 +1687,6 @@ void test_prvProcessEthernetPacket_IPv4FrameType_AptData( void )
     prvProcessEthernetPacket( pxNetworkBuffer );
 }
 
-void test_xIsIPv4Multicast_NotMultiCast( void )
-{
-    BaseType_t xReturn;
-    uint32_t ulIPAddress = 0;
-
-    xReturn = xIsIPv4Multicast( ulIPAddress );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xReturn );
-}
-
-void test_xIsIPv4Multicast_NotMultiCast2( void )
-{
-    BaseType_t xReturn;
-    uint32_t ulIPAddress = FreeRTOS_htonl( 0xF0000000 );
-
-    xReturn = xIsIPv4Multicast( ulIPAddress );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xReturn );
-}
-
-void test_xIsIPv4Multicast_IsMultiCast( void )
-{
-    BaseType_t xReturn;
-    uint32_t ulIPAddress = FreeRTOS_htonl( 0xF0000000 - 1 );
-
-    xReturn = xIsIPv4Multicast( ulIPAddress );
-
-    TEST_ASSERT_EQUAL( pdTRUE, xReturn );
-}
-
-void test_prvAllowIPPacketIPv4( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_FragmentedPacket( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    pxIPHeader->usFragmentOffset = ipFRAGMENT_OFFSET_BIT_MASK;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_FragmentedPacket1( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    pxIPHeader->usFragmentOffset = ipFRAGMENT_FLAGS_MORE_FRAGMENTS;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_IncorrectLength( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    pxIPHeader->ucVersionHeaderLength = 0xFF;
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_NotMatchingIP( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxNetworkBuffer->pxEndPoint = NULL;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xAB12CD34;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER + 1;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SourceIPBrdCast_DestIPMatch( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xAB12CD34;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    pxIPHeader->ulSourceIPAddress = 0xFFFFFFFF;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SourceIPBrdCast_DestIPBrdCast( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xAB12CD34;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = 0xFFFFFFFF;
-
-    pxIPHeader->ulSourceIPAddress = 0xFFFFFFFF;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SourceIPBrdCast_DestIPBrdcast1( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    NetworkEndPoint_t xEndPoints;
-    IPV4Parameters_t * xIPv4Addressing = &( xEndPoints.ipv4_settings );
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xAB12CD34;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = xIPv4Addressing->ulBroadcastAddress;
-
-    pxIPHeader->ulSourceIPAddress = 0xFFFFFFFF;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SourceIPBrdCast_DestIPLLMNR( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xAB12CD34;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = ipLLMNR_IP_ADDR;
-
-    pxIPHeader->ulSourceIPAddress = 0xFFFFFFFF;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SourceIPBrdCast_NoLocalIP( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0x00;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER + 1;
-
-    pxIPHeader->ulSourceIPAddress = 0xFFFFFFFF;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_DestMACBrdCast_DestIPUnicast( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0x00;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = 0x00;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SrcMACBrdCast( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xSourceAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SrcMACBrdCast2( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xSourceAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_SrcIPAddrIsMulticast( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    memset( pxNetworkBuffer, 1, sizeof( NetworkBufferDescriptor_t ) );
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-    pxNetworkBuffer->pxEndPoint = NULL;
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = FreeRTOS_htonl( 0xE0000000 + 1 );
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_IncorrectChecksum( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer );
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    FreeRTOS_FindEndPointOnMAC_ExpectAnyArgsAndReturn( NULL );
-
-    usGenerateChecksum_ExpectAndReturn( 0U, ( uint8_t * ) &( pxIPHeader->ucVersionHeaderLength ), ( size_t ) uxHeaderLength, ipCORRECT_CRC - 1 );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_IncorrectProtocolChecksum( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    FreeRTOS_FindEndPointOnMAC_ExpectAnyArgsAndReturn( NULL );
-
-    usGenerateChecksum_ExpectAndReturn( 0U, ( uint8_t * ) &( pxIPHeader->ucVersionHeaderLength ), ( size_t ) uxHeaderLength, ipCORRECT_CRC );
-
-    usGenerateProtocolChecksum_ExpectAndReturn( ( uint8_t * ) ( pxNetworkBuffer->pucEthernetBuffer ), pxNetworkBuffer->xDataLength, pdFALSE, ( uint16_t ) ( ipCORRECT_CRC + 1 ) );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eReleaseBuffer, eResult );
-}
-
-void test_prvAllowIPPacketIPv4_HappyPath( void )
-{
-    eFrameProcessingResult_t eResult;
-    IPPacket_t * pxIPPacket;
-    NetworkBufferDescriptor_t * pxNetworkBuffer, xNetworkBuffer;
-    UBaseType_t uxHeaderLength = 0;
-    uint8_t ucEthBuffer[ ipconfigTCP_MSS ];
-    IPHeader_t * pxIPHeader;
-
-    memset( ucEthBuffer, 0, ipconfigTCP_MSS );
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthBuffer;
-    pxIPPacket = ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer;
-    pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xFFFFFFFF;
-
-    pxIPHeader->ucVersionHeaderLength = 0x45;
-
-    pxIPHeader->ulDestinationIPAddress = *ipLOCAL_IP_ADDRESS_POINTER;
-
-    memcpy( pxIPPacket->xEthernetHeader.xDestinationAddress.ucBytes, xBroadcastMACAddress.ucBytes, sizeof( MACAddress_t ) );
-
-    pxIPHeader->ulSourceIPAddress = 0xC0C00101;
-
-    FreeRTOS_FindEndPointOnIP_IPv4_ExpectAnyArgsAndReturn( NULL );
-
-    FreeRTOS_FindEndPointOnMAC_ExpectAnyArgsAndReturn( NULL );
-
-    usGenerateChecksum_ExpectAndReturn( 0U, ( uint8_t * ) &( pxIPHeader->ucVersionHeaderLength ), ( size_t ) uxHeaderLength, ipCORRECT_CRC );
-
-    usGenerateProtocolChecksum_ExpectAndReturn( ( uint8_t * ) ( pxNetworkBuffer->pucEthernetBuffer ), pxNetworkBuffer->xDataLength, pdFALSE, ipCORRECT_CRC );
-
-    eResult = prvAllowIPPacketIPv4( pxIPPacket, pxNetworkBuffer, uxHeaderLength );
-
-    TEST_ASSERT_EQUAL( eProcessBuffer, eResult );
-}
-
 void test_prvProcessIPPacket_HeaderLengthSmaller( void )
 {
     eFrameProcessingResult_t eResult;
